chore(rust): 'not_unsafe_ptr_arg_deref' lints
these functions are /technically/ unsafe with respect to pointer
argument dereferencing, which is to be expected from FFI bridge code,
and as such 'clippy' throws an error if they aren't marked as such & a
warning if the 'unsafe' function doesn't have associated safety docs.

it's possible that some of these are false positives[0] that could be
addressed with `#[allow(clippy::not_unsafe_ptr_arg_deref)]`; I've added
'TODO: [publish-crate]' to all of these functions as a reminder to go
through and review them if/when this is published to 'crates.io'.

[0]: https://github.com/rust-lang/rust-clippy/issues/3045