fix(core): CRPCError `details` field is a pointer
in Rust, the `details` field of `CRPCError` is a pointer to a `CArray`,
but in Haskell we were attempting to unmarshall it directly.

the `CRPCError` type now accurately represents this as a
`Ptr (CArray a)`, and `CArray Word8 -> ByteString` code has been updated
to correctly dereference this before copying the data over.

nb. prior to this change, the `size` field must have just been some
    unspecified chunk of memory, which makes me wonder where some of the
    displayed values we were seeing in CI actually came from.