feat(sdk): provide time-skipping server to sdk
this is a little messy, but we need to make the time-skipping server
available to the SDK's test suite so we can actually run tests against
it.

that means it needs to be provided to the Nix derivation, rather than
just being made available to the development shell.

unfortunately this means our test matrix needs to be pared back, as
`aarch64-linux` test server binaries are not (yet) published and the
native build environment for them is difficult to wrangle.