fix(nix): patch time-skipping server ELF headers
without this the test fails in CI with a cryptic 'file not found' error.

when run interactively — in the dev shell — on an x86-64 Linux host, the
actual error message indicates that the target was built on a generic
Linux host and is incompatible with NixOS.

this should be resolved by calling `patchelf`.