fix(nix): guard time-skipping server on aarch64
this should allow the project to build on 'aarch64-linux' by selectively
avoiding the 'temporal-test-server' derivation on that platform.

unfortunately the test suite will fail if 'doCheck = true', but if we
_really_ want to support that we can add a flag to the 'temporal-sdk'
'package.yaml' file to ignore time-skipping tests on that platform.

now that we have integration tests that use the time-skipping server,
we also need to guard against those being run on 'aarch64-linux' with
CPP in Haskell; otherwise these would fail, being unable to find the
executable.

all of this will go away as soon as we can get an 'aarch64-linux' binary
for the test server uploaded.