initial support for time-skipping (#126)
* initial support for time-skipping

* not backwards though

* feat(sdk): provide time-skipping server to sdk

this is a little messy, but we need to make the time-skipping server
available to the SDK's test suite so we can actually run tests against
it.

that means it needs to be provided to the Nix derivation, rather than
just being made available to the development shell.

unfortunately this means our test matrix needs to be pared back, as
`aarch64-linux` test server binaries are not (yet) published and the
native build environment for them is difficult to wrangle.

* small simplification

* test child workflows too

* fix(nix): patch time-skipping server ELF headers

without this the test fails in CI with a cryptic 'file not found' error.

when run interactively — in the dev shell — on an x86-64 Linux host, the
actual error message indicates that the target was built on a generic
Linux host and is incompatible with NixOS.

this should be resolved by calling `patchelf`.

* add TestService functions in the sdk

* fix(nix): guard time-skipping server on aarch64

this should allow the project to build on 'aarch64-linux' by selectively
avoiding the 'temporal-test-server' derivation on that platform.

unfortunately the test suite will fail if 'doCheck = true', but if we
_really_ want to support that we can add a flag to the 'temporal-sdk'
'package.yaml' file to ignore time-skipping tests on that platform.

now that we have integration tests that use the time-skipping server,
we also need to guard against those being run on 'aarch64-linux' with
CPP in Haskell; otherwise these would fail, being unable to find the
executable.

all of this will go away as soon as we can get an 'aarch64-linux' binary
for the test server uploaded.

---------

Co-authored-by: jkachmar <j@mercury.com>