chore(deps): update temporal-cli 1.3.0 -> 1.5.1
this won't be necessary whenever we update our nixpkgs pin, but for now
it's not too much extra work to vendor a derivation that pulls in the
precompiled CLI binary.

since this isn't something that's /strictly/ necessary for build/test,
it's going in a separate 'development' overlay so people don't
implicitly pull it in if they just want the native bridge library &
test server.